Remove restart test
This test is not applicable anymore now that containers are not stopped
when the daemon is restored.
